Splits x86 JIT's IR handling over to a jump table
This is changes the behaviour to be exactly like the ARM64 side.
No behaviour change beyond that.
